feat: swagger

+++ to-do/src/main/java/com/todolist/todo/services/ToDoService.java
@@ -2,6 +2,7 @@ package com.todolist.todo.services;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.UUID;
 import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
@@ -13,27 +14,47 @@ import com.todolist.todo.dtos.GetToDoResponseDto;
 import com.todolist.todo.dtos.UpdateToDoDto;
 import com.todolist.todo.exceptions.BadRequestException;
 import com.todolist.todo.repositories.ToDoRepository;
+import com.todolist.todo.repositories.UserRepository;
 
 import lombok.RequiredArgsConstructor;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.BeanUtils;
-import java.time.LocalDateTime;
-import java.sql.Timestamp;
+import org.springframework.beans.factory.annotation.Value;
+
+import com.auth0.jwt.JWT;
+import com.auth0.jwt.algorithms.Algorithm;
+import com.auth0.jwt.interfaces.DecodedJWT;
 
 @Service
 @RequiredArgsConstructor
 public class ToDoService {
   private final ToDoRepository toDoRepository;
+  private final UserRepository UserRepository;
   private final ModelMapper modelMapper = new ModelMapper();
+  @Value("${api.security.token.secret}")
+  private String secretKey;
 
-  public GetToDoResponseDto create(CreateToDoDto body, UserModel user) {
+  private UserModel getUserFromToken(String token) {
+
+    DecodedJWT decodedJWT = JWT.require(Algorithm.HMAC256(secretKey))
+        .build()
+        .verify(token.replace("Bearer ", ""));
+
+    String userId = decodedJWT.getSubject();
+
+    Optional<UserModel> user = UserRepository.findById(UUID.fromString(userId));
+
+    return user.get();
+  }
+
+  public GetToDoResponseDto createToDo(CreateToDoDto body, String token) {
+
+    UserModel userLogged = this.getUserFromToken(token);
 
     var toDo = new ToDoModel();
     BeanUtils.copyProperties(body, toDo);
-    toDo.setUser(user);
-    toDo.setCreatedAt(Timestamp.valueOf(LocalDateTime.now()));
-    toDo.setUpdatedAt(Timestamp.valueOf(LocalDateTime.now()));
+    toDo.setUser(userLogged);
     toDoRepository.save(toDo);
 
     GetToDoResponseDto toDoFormated = modelMapper.map(
@@ -42,20 +63,30 @@ public class ToDoService {
     return toDoFormated;
   }
 
-  public List<GetToDoResponseDto> listAll() {
+  public List<GetToDoResponseDto> listAllToDos(String token) {
+
+    UserModel userLogged = this.getUserFromToken(token);
 
     List<ToDoModel> allToDos = toDoRepository.findAll();
 
-    return allToDos.stream()
-        .map(user -> modelMapper.map(user, GetToDoResponseDto.class))
+    List<GetToDoResponseDto> filteredToDos = allToDos.stream()
+        .filter(toDo -> toDo.getUser().getId().equals(userLogged.getId()))
+        .map(toDo -> modelMapper.map(toDo, GetToDoResponseDto.class))
         .collect(Collectors.toList());
+
+    return filteredToDos;
   }
 
-  public GetToDoResponseDto list(Long id) {
-    Optional<ToDoModel> foundToDo = toDoRepository.findById(id);
+  public GetToDoResponseDto listToDo(Long id, String token) {
 
+    Optional<ToDoModel> foundToDo = toDoRepository.findById(id);
     if (foundToDo.isEmpty()) {
       throw new BadRequestException("To do %d does not exist!".formatted(id));
+    }
+
+    UserModel userLogged = this.getUserFromToken(token);
+    if (foundToDo.get().getUser() != userLogged) {
+      throw new BadRequestException("You can only list your to dos");
     }
 
     GetToDoResponseDto toDoFormated = modelMapper.map(
@@ -64,11 +95,16 @@ public class ToDoService {
     return toDoFormated;
   }
 
-  public GetToDoResponseDto update(Long id, UpdateToDoDto body) {
+  public GetToDoResponseDto updateToDo(Long id, UpdateToDoDto body, String token) {
 
     Optional<ToDoModel> foundToDo = toDoRepository.findById(id);
     if (foundToDo.isEmpty()) {
       throw new BadRequestException("To do %d does not exist!".formatted(id));
+    }
+
+    UserModel userLogged = this.getUserFromToken(token);
+    if (foundToDo.get().getUser() != userLogged) {
+      throw new BadRequestException("You can only update your to dos");
     }
 
     var toDo = foundToDo.get();
@@ -89,8 +125,6 @@ public class ToDoService {
       toDo.setPriority(body.priority());
     }
 
-    toDo.setUpdatedAt(Timestamp.valueOf(LocalDateTime.now()));
-
     toDoRepository.save(toDo);
 
     GetToDoResponseDto toDoFormated = modelMapper.map(
@@ -99,11 +133,17 @@ public class ToDoService {
     return toDoFormated;
   }
 
-  public void deleteToDo(Long id) {
+  public void deleteToDo(Long id, String token) {
     Optional<ToDoModel> foundToDo = toDoRepository.findById(id);
     if (foundToDo.isEmpty()) {
       throw new BadRequestException("This user does not exist");
     }
+
+    UserModel userLogged = this.getUserFromToken(token);
+    if (foundToDo.get().getUser() != userLogged) {
+      throw new BadRequestException("You can only update your to dos");
+    }
+
     toDoRepository.delete(foundToDo.get());
   }
 }
